transport: fix issues with BAP volumes > 127
Change types used for volumes to int, for BAP volumes > 127 and to
allow -1 as it is used to mean no volume.

Fix BAP max volume check in transport.

Fixes setting BAP volumes.

+++ profiles/audio/transport.h
@@ -22,14 +22,14 @@ const char *media_transport_get_path(struct media_transport *transport);
 void *media_transport_get_stream(struct media_transport *transport);
 struct btd_device *media_transport_get_dev(struct media_transport *transport);
 int media_transport_get_volume(struct media_transport *transport,
-					int8_t *volume);
+					int *volume);
 void media_transport_update_delay(struct media_transport *transport,
 							uint16_t delay);
 void media_transport_update_volume(struct media_transport *transport,
-								int8_t volume);
+								int volume);
 void transport_get_properties(struct media_transport *transport,
 							DBusMessageIter *iter);
 
-int8_t media_transport_get_device_volume(struct btd_device *dev);
+int media_transport_get_device_volume(struct btd_device *dev);
 void media_transport_update_device_volume(struct btd_device *dev,
-								int8_t volume);
+								int volume);
